% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlaststops.R
\name{getlaststops}
\alias{getlaststops}
\title{Create a stops table for makexmlsiri}
\usage{
getlaststops(GTFSroutes., GTFStrips., GTFSstop_times., GTFSstops.)
}
\arguments{
\item{GTFSroutes.}{A GTFS routes table, best load into environment with \code{\link{readGTFS}}}

\item{GTFStrips.}{A GTFS trips table, best load into environment with \code{\link{readGTFS}}}

\item{GTFSstop_times.}{A GTFS stop_times table, best load into environment with \code{\link{readGTFS}}}

\item{GTFSstops.}{A GTFS stops table, best load into environment with \code{\link{readGTFS}}}
}
\value{
A \code{\link[base]{data.frame}} containing for each unique route_id in the GTFS tables it's
shape_id and the stop_id and stop_code of the last stop in that shape
}
\description{
Creates a
}
\details{
Uses SQL within R to create a table containing each unique route_id in the GTFS provided,
for each route, it's shape_id and the stop_code and stop_id of the last stop in that shape.
usefull for the creation of XML files using \code{\link[SIRItoGTFS]{makexmlsiri}}
}
\seealso{
\code{\link{SIRItoGTFS}}, \code{\link[SIRItoGTFS]{makexmlsiri}}
}
\keyword{misc}
\keyword{utilities}
