% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makexmlsiri.R
\name{makexmlsiri}
\alias{makexmlsiri}
\title{Create the required XML files for SIRI downloads}
\usage{
makexmlsiri(
  stop_codes,
  route_codes,
  wd = NULL,
  xml_size = 200,
  RequestorRef = NULL,
  xml_name = "xml",
  newfolder = TRUE
)
}
\arguments{
\item{stop_codes}{A \code{\link[base]{vector} of stop_code integers taken from the GTFS tables}. the XML will not fail to create if there are no integers.}

\item{route_codes}{A \code{\link[base]{vector} of route_id integers taken from the GTFS tables}. the XML will not fail to create if there are no integers.}

\item{wd}{the working directory to write the files in, if newfolder is TRUE a new folder called 'xml' will be created, if there allready is a folder called 'xml' in the working directory a prompt window will ask you to choose a new folder name.}

\item{xml_size}{Default is 200, the number of stops written to each XML file.}

\item{RequestorRef}{A unique Requestor key supplied by the Israeli MOT, it is needed since the API requires registration and a static IP address.}

\item{xml_name}{the prefix to be used for the new XML files. default is 'xml'.}

\item{newfolder}{Create a new folder for the files?}
}
\value{
Returns a string message, and files written to working directory.
}
\description{
Creates XML files used for sending SIRI requests to Israel's MOT API.
}
\details{
Creates XML files used for sending SIRI requests to Israel's MOT API.
note that an API key "RequestorRef" is needed for the process.
It is recommended to use a single sorted table containing both stop_codes and route_ids.
}
\seealso{
\code{\link{SIRItoGTFS}}
}
\keyword{misc}
\keyword{utilities}
