% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlaststops.R
\name{getlaststops}
\alias{getlaststops}
\title{Create a stops table for makexmlsiri}
\usage{
getlaststops(GTFSroutes. = GTFSroutes, GTFStrips. = GTFStrips,
  GTFSstop_times. = GTFSstop_times, GTFSstops. = GTFSstops)
}
\arguments{
\item{GTFSroutes.}{A GTFS routes table, best load into environment with \code{\link{readGTFS}}}

\item{GTFStrips.}{A GTFS trips table, best load into environment with \code{\link{readGTFS}}}

\item{GTFSstop_times.}{A GTFS stop_times table, best load into environment with \code{\link{readGTFS}}}

\item{GTFSstops.}{A GTFS stops table, best load into environment with \code{\link{readGTFS}}}
}
\value{
A \code{\link[base]{data.frame}} containing for each unique route_id in the GTFS tables it's
shape_id and the stop_id and stop_code of the last stop in that shape
}
\description{
Creates a
}
\keyword{misc}
\keyword{utilities}
